requirements.txt
python-dotenv
requests
pandas
matplotlib
py-clob-client
polymarket-apis
